Tested commit of changes so far.
