+++ src/Model/CohortMerge.h
@@ -3,7 +3,7 @@
 
 #include "Cohort.h"
 #include "Pear.h"
-#include "MadingleyModelInitialisation.h"
+#include "MadingleyInitialisation.h"
 #include "NonStaticSimpleRNG.h"
 #include "Parameters.h"
 #include <set>
@@ -24,7 +24,7 @@ public:
     @param TotalNumberOfCohorts The total number of cohorts in this grid cell 
     @param TargetCohortThreshold The target threshold to reduce the number of cohorts to 
     @return The number of cohorts that have been merged */
-    int MergeToReachThresholdFast( GridCell&, MadingleyModelInitialisation& );
+    int MergeToReachThresholdFast( GridCell&, MadingleyInitialisation& );
     
     /** \brief An instance of simple random number generator class */
     NonStaticSimpleRNG mRandomNumber;
