+++ src/Model/Metabolism.h
@@ -1,60 +1,28 @@
-#ifndef METABOLISM_H
-#define METABOLISM_H
-#include <EcologicalProcessWithinGridCell.h>
-#include <MetabolismImplementation.h>
-#include <TMetabolismHeterotroph.h>
-#include <TMetabolismEndotherm.h>
-#include <TMetabolismEctotherm.h>
-#include <ThreadLocked.h>
-/** \file Metabolism.h
- * \brief the Metabolism header file
- */
+#ifndef METABOLISM
+#define METABOLISM
 
-//
-//namespace Madingley
-//{
+#include "EcologicalProcessWithinGridCell.h"
+#include "MetabolismImplementation.h"
+#include "TMetabolismHeterotroph.h"
+#include "TMetabolismEndotherm.h"
+#include "TMetabolismEctotherm.h"
+#include "ThreadLocked.h"
 
 /** \brief  Performs metabolism */
-class Metabolism: public EcologicalProcessWithinGridCell {
+class Metabolism : public EcologicalProcessWithinGridCell {
 public:
-    /** \brief The available implementations of the metabolism process*/
-    //----------------------------------------------------------------------------------------------
-    //Variables
-    //----------------------------------------------------------------------------------------------
-    map<string, MetabolismImplementation*> Implementations;
-    //----------------------------------------------------------------------------------------------
-    //Methods
-    //----------------------------------------------------------------------------------------------
-
-    //----------------------------------------------------------------------------------------------
-
     /** \brief Constructor Metabolism: fills the list of available implementations of metabolism*/
-    Metabolism( string globalModelTimeStepUnit ) {
-        // Add the basic endotherm metabolism implementation to the list of implementations
-        MetabolismEndotherm* MetabolismEndothermImplementation = new MetabolismEndotherm( globalModelTimeStepUnit );
-        Implementations["basic endotherm"] = MetabolismEndothermImplementation;
+    Metabolism( std::string globalModelTimeStepUnit );
 
-        // Add the basic ectotherm metabolism implementation to the list of implementations
-        MetabolismEctotherm* MetabolismEctothermImplementation = new MetabolismEctotherm( globalModelTimeStepUnit );
-        Implementations["basic ectotherm"] = MetabolismEctothermImplementation;
-    }
-    //----------------------------------------------------------------------------------------------
+    /** Destrcutor to tidy up pointers */
+    ~Metabolism( );
 
-    /** Destrcutor to tidy up pointers*/
-    ~Metabolism( ) {
-        delete Implementations["basic endotherm"];
-        delete Implementations["basic ectotherm"];
-    }
-    //----------------------------------------------------------------------------------------------
-
+private:
     /** \brief Initializes an implementation of metabolism
     @param gcl The current grid cell 
     @param params A bunch of parameters 'n' stuff 'n' things 
     @param implementationKey The name of the implementation of metabolism to initialize  */
-    void InitializeEcologicalProcess( GridCell& gcl, MadingleyModelInitialisation& params, string implementationKey ) {
-
-    }
-    //----------------------------------------------------------------------------------------------
+    void InitializeEcologicalProcess( GridCell&, MadingleyInitialisation&, std::string );
 
     /** \brief Run metabolism
     @param gcl The current grid cell 
@@ -63,24 +31,7 @@ public:
     @param partial Thread-locked variables 
     @param currentMonth The current model month
     @param params some parameters  */
-    void RunEcologicalProcess( GridCell& gcl,
-            Cohort& actingCohort,
-            unsigned currentTimestep,
-            ThreadLockedParallelVariables& partial,
-            unsigned currentMonth, MadingleyModelInitialisation& params ) {
-
-        if( params.mCohortFunctionalGroupDefinitions.GetTraitNames( "Heterotroph/Autotroph", actingCohort.mFunctionalGroupIndex ) == "heterotroph" ) {
-            if( params.mCohortFunctionalGroupDefinitions.GetTraitNames( "Endo/Ectotherm", actingCohort.mFunctionalGroupIndex ) == "endotherm" ) {
-
-                Implementations["basic endotherm"]->RunMetabolism( actingCohort, currentTimestep, currentMonth );
-            } else {
-                Implementations["basic ectotherm"]->RunMetabolism( actingCohort, currentTimestep, currentMonth );
-
-            }
-
-        }
-
-    }
-    //----------------------------------------------------------------------------------------------
+    void RunEcologicalProcess( GridCell&, Cohort&, unsigned, ThreadLockedParallelVariables&, unsigned, MadingleyInitialisation& );
+    std::map< std::string, MetabolismImplementation* > Implementations;
 };
 #endif
