+++ src/Model/TRevisedHerbivory.h
@@ -91,7 +91,7 @@ public:
     @param params The definitions of model parameters 
     \remark This only works if: a) herbivory is initialised in every grid cell; and b) if parallelisation is done by latitudinal strips
     It is critical to run this every time step */
-    void InitializeEatingPerTimeStep( GridCell& gcl, MadingleyModelInitialisation& params ) {
+    void InitializeEatingPerTimeStep( GridCell& gcl, MadingleyInitialisation& params ) {
         // Store the specified cell area in this instance of this herbivory implementation
         CellArea = gcl.GetCellArea( );
         CellAreaHectares = CellArea * 100;
@@ -104,7 +104,7 @@ public:
     @param gcl The current grid cell 
     @param actingCohort The acting cohort 
     @param params The definitions for stuff in the model */
-    void GetEatingPotentialTerrestrial( GridCell& gcl, Cohort& actingCohort, MadingleyModelInitialisation& params ) {
+    void GetEatingPotentialTerrestrial( GridCell& gcl, Cohort& actingCohort, MadingleyInitialisation& params ) {
         // Set the total biomass eaten by the acting cohort to zero
         mTotalBiomassEatenByCohort = 0.0;
 
@@ -147,7 +147,7 @@ public:
     @param gcl The current grid cell 
     @param actingCohort The acting cohort 
     @params All your base are belong to us */
-    void GetEatingPotentialMarine( GridCell& gcl, Cohort& actingCohort, MadingleyModelInitialisation& params ) {
+    void GetEatingPotentialMarine( GridCell& gcl, Cohort& actingCohort, MadingleyInitialisation& params ) {
         // Set the total biomass eaten by the acting cohort to zero
         mTotalBiomassEatenByCohort = 0.0;
 
@@ -191,7 +191,7 @@ public:
     @param params  The model parameters */
     void RunEating( GridCell& gcl, Cohort& actingCohort,
             unsigned currentTimestep,
-            MadingleyModelInitialisation& params ) {
+            MadingleyInitialisation& params ) {
 
         EdibleScaling = 1.0;
         if( !gcl.IsMarine( ) ) EdibleScaling = 0.1;
